MAGECLOUD-2851: Throw exception of json decode errors
